Refactor: 도미노 및 프로젝트 스키마 책임 분리
#11
